bugfix for problem when putting batch containing PUT requests whith identical version combined with PUT requests with changes

+++ sql/schema.sql
@@ -169,7 +169,7 @@ CREATE TABLE "personrelations" (
 CREATE TABLE "cities" (
   "key" int4 unique,
   "name" text,
-  "nisCode" numeric,
+  "nisCode" int,
   "$$meta.deleted" boolean default false,
   "$$meta.modified" timestamp with time zone not null default current_timestamp,
   "$$meta.created" timestamp with time zone not null default current_timestamp
